fileservice: move objcache/lruobjcache to lrucache
