+++ pkg/vm/engine/tae/model/index.go
@@ -18,7 +18,7 @@ import (
 	"context"
 
 	"github.com/matrixorigin/matrixone/pkg/fileservice"
-	"github.com/matrixorigin/matrixone/pkg/fileservice/objcache/lruobjcache"
+	"github.com/matrixorigin/matrixone/pkg/fileservice/lrucache"
 )
 
 type LRUCache interface {
@@ -28,12 +28,12 @@ type LRUCache interface {
 }
 
 type simpleLRU struct {
-	impl lruobjcache.LRU[any, fileservice.Bytes]
+	impl lrucache.LRU[any, fileservice.Bytes]
 }
 
 func NewSimpleLRU(capacity int64) LRUCache {
 	return &simpleLRU{
-		impl: *lruobjcache.New[any, fileservice.Bytes](capacity, nil, nil),
+		impl: *lrucache.New[any, fileservice.Bytes](capacity, nil, nil),
 	}
 }
